[SIDE-1196] Refactoring
